+++ Next2/src/Next2/Models/GiftCardModel.cs
@@ -13,7 +13,7 @@ namespace Next2.Models
 
         public int GiftCardNumber { get; set; }
 
-        public float GiftCardFunds { get; set; }
+        public decimal GiftCardFunds { get; set; }
 
         public bool IsRegistered { get; set; }
     }
