+++ Next2/src/Next2/Models/SetModel.cs
@@ -15,9 +15,9 @@ namespace Next2.Models
 
         public string Title { get; set; }
 
-        public float Price { get; set; }
+        public decimal Price { get; set; }
 
-        public float Bonus { get; set; }
+        public decimal Bonus { get; set; }
 
         public string ImagePath { get; set; }
     }
