+++ Next2/src/Next2/ViewModels/Mobile/TipsPageViewModel.cs
@@ -90,7 +90,7 @@ namespace Next2.ViewModels.Mobile
         {
             await _popupNavigation.PopAsync();
 
-            if (parameters.TryGetValue(Constants.DialogParameterKeys.TIP_VALUE_DIALOG, out float value))
+            if (parameters.TryGetValue(Constants.DialogParameterKeys.TIP_VALUE_DIALOG, out decimal value))
             {
                 SelectedTipItem.Value = value;
                 SelectedTipItem.Text = LocalizationResourceManager.Current["CurrencySign"] + $" {value:F2}";
